FEDEV-2190 map trade history errors
