+++ src/components/TradeHistory/TradeHistoryRow/utils/position.ts
@@ -246,7 +246,7 @@ export const formatPositionMessage = (
       actionComment:
         error &&
         lines({
-          text: getErrorTooltipTitle(error.name, true),
+          text: getErrorTooltipTitle(error.name, true, error.args),
           state: "error",
         }),
       size: customSize,
@@ -272,7 +272,7 @@ export const formatPositionMessage = (
       const error = tradeAction.reasonBytes ? tryDecodeCustomError(tradeAction.reasonBytes) ?? undefined : undefined;
       const errorComment = error
         ? lines({
-            text: getErrorTooltipTitle(error.name, false),
+            text: getErrorTooltipTitle(error.name, false, error.args),
             state: "error",
           })
         : undefined;
@@ -334,7 +334,7 @@ export const formatPositionMessage = (
       actionComment:
         error &&
         lines({
-          text: getErrorTooltipTitle(error.name, false),
+          text: getErrorTooltipTitle(error.name, false, error.args),
           state: "error",
         }),
       priceComment: lines(
@@ -396,7 +396,7 @@ export const formatPositionMessage = (
       actionComment:
         error &&
         lines({
-          text: getErrorTooltipTitle(error.name, true),
+          text: getErrorTooltipTitle(error.name, true, error.args),
           state: "error",
         }),
       size: customSize,
@@ -455,7 +455,7 @@ export const formatPositionMessage = (
       actionComment:
         error &&
         lines({
-          text: getErrorTooltipTitle(error.name, false),
+          text: getErrorTooltipTitle(error.name, false, error.args),
           state: "error",
         }),
       priceComment: lines(
@@ -512,7 +512,7 @@ export const formatPositionMessage = (
       actionComment:
         error &&
         lines({
-          text: getErrorTooltipTitle(error.name, false),
+          text: getErrorTooltipTitle(error.name, false, error.args),
           state: "error",
         }),
       priceComment: lines(
